Show an error if there’s a file with no extension
This aims to help catch a common error: people adding a markdown file for a post but forgetting to add the `.md` file extension.

+++ eleventy.config.js
@@ -2,6 +2,8 @@ import fs from 'node:fs/promises'
 
 import { nhsukEleventyPlugin } from '@x-govuk/nhsuk-eleventy-plugin'
 
+import { findFileWithoutExtension } from './lib/find-file-without-extension.js'
+
 const serviceName = 'Digital prevention services design history'
 
 export default function (eleventyConfig) {
@@ -174,6 +176,17 @@ export default function (eleventyConfig) {
 
   // Reset contents of output directory before each build
   eleventyConfig.on('eleventy.before', async ({ directories, runMode }) => {
+    // Check for files without extensions in the app directory
+    const fileWithoutExtension = await findFileWithoutExtension(
+      directories.input
+    )
+
+    if (fileWithoutExtension) {
+      throw new Error(
+        `Found file without extension: ${fileWithoutExtension}\n\nDid you forget to add an .md to it?`
+      )
+    }
+
     if (runMode === 'build') {
       await fs.rm(directories.output, {
         force: true,
